fix: mobile wallet init using initialization context
